- Added maintainer notes to code. - Fixed small bugs within `mps_sequential.py`
